server/requirements.txt
setuptools
#==58.1.0
wheel
#==0.37.1
django
#==4.0.3
djangorestframework
#==3.13.1
django-filter
#==21.1
django-cors-headers
#==3.11.0
djangorestframework-simplejwt
#==5.1.0
drf-yasg
#==1.20.0
python-dotenv
#==0.20.0
pillow
#==9.1.0
psycopg2-binary
#==2.9.3
django-filebrowser
#==4.0.1
django-tinymce
#==3.4.0
gspread
#==5.3.2
oauth2client
#==4.1.3
pyjwt
#==2.3.0
filebrowser #==1.1.3
django-grappelli
drf_yasg
